main.dart created
I want to try and get one page, or one section done each day. That way I can keep slowly working on it without just exploding and getting close but not really completing it. The Half app was basically completed, I just needed to figure out how to put it on the Play Store, add a logo and stuff.