root.dart
I think that I about got the new path system set up.

I think that overall this will be a lot easier to manage and get a head around.

Next up I need to finalize the user interfaces for login.dart, welcome.dart, forgotPassword.dart.

After that, I need to start working on the home.dart page and getting all of the data stuff setup for it.