Route navigation update
I am trying to get a new route navigation system setup. I want something that is easy to use and can be expanded in the future, so I am thinking about going with something similar to the cloudFirestore.dart file layout

I think that I am making some good headway on it, but I still have quite a bit of work to do along with quite a bit of testing.

I think my next step is going to see if I can create a method for navigating at the creation of the app (which I may have to go ahead and code the WaitingScreen() widget). If I can get that done, then I think that this navigation system will be complete.

I will still have to add the other pages and change the name of the file though