animationStandards.dart && routeNavigation.dart update
I have done quite a bit of updating.

I changed the routeNavigation.dart design a little. Now each method will have the ability to either send a pageRoute or the actual widget itself. I may change the order of the two, but other than that I think that it is done