login.dart update
Trying to get this page setup now.

I want to get a few pages setup and then go in and see if it is actually working. I have not tested this code in like 3 weeks now ':D so we are going to see how it goes. I think that if I do the login.dart page I could test it, but I may go ahead and do the intro.dart page too. That will leave me with 2 pages left in the app to do.

Once I get the login.dart and intro.dart pages complete, I'm going to test it and then probably spend another few days perfecting them.

I want to implement error toast messages for signIn and signUp. Then I also want to get the password visibility button working and positioned correctly.