routeNavigation.dart
Okay I think that I am getting this closer. I think that the last thing I have to do, in order for it to at least run, is to finish putting together the main.dart page. If I do this, then I think that the pages and widgets will be able to call for text translations.