home.dart & detail.dart update
HUGE UPDATE!

home.dart has been updated with a carousel for cards and taps for the cards. I need to:
- Decrease distance between cards
- All the other things previously stated

detail.dart has been created and worked on for a bit. It now has a hero animation going between the home.dart page and the detail.dart page. The tag for each card is 'imageCard$imageURL' so that each one of them is unique