home.dart update
Working on the grid view currently. I want to try and see if I can create it to where both the list and the grid use the same FutureBuilder, that way we aren't having to redo it all the time. This might not actually be a problem if we are simply using the future builder to reset the already initialized list<DocumentSnapshot>. This might also cause a problem because we might not be able to order the grid how we want it if we are sending everything in one at a time, we might want to create a function to break up the list into different sublists based on the month that they have in their id (or we could just create a new variable and if the new month is different then we add in a new header and go on reading in the next ones)